Use jQuery to switch game modes and start adding level select/world map.
Use jQuery to slightly simplify code for switching between game modes,
e.g. the shop, the world map, and combat.

Also, begin adding a level select screen/world map. Right now, the
levels are organizing like a tree, and the user can only see and play
levels that they've already beaten, and levels that are immediate
children of level's they've already beaten. To keep track of which
levels the user has beaten, a User object is added (in
GameCode/User.js), which holds how much gold the user has in addition to
which level's they've beaten, and can be saved to/loaded from
localStorage.

Right now, the only difference between levels is their difficulty level
and how much gold they reward. The difficulty and gold are stored in the
HTML for the world map. The HTML for the levels also determines their
hiearchy (i.e. if a .level element is a child of another .level element,
then the user needs to beat the parent level before they can see the
child level).